refactor: 에러 코드 적용
