+++ src/main/java/com/example/board/dto/response/ApiResponse.java
@@ -11,6 +11,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class ApiResponse<T> {
+
     private Boolean isSuccess;
     private int statusCode;
     private T data;
