+++ src/main/java/com/example/board/service/PostService.java
@@ -7,15 +7,15 @@ import com.example.board.dto.request.CreatePostRequest;
 import com.example.board.dto.request.PostSearchCondition;
 import com.example.board.dto.request.UpdatePostRequest;
 import com.example.board.dto.response.PostResponse;
-import com.example.board.repository.PostRepository;
-import com.example.board.repository.UserRepository;
+import com.example.board.exception.CustomException;
+import com.example.board.exception.ErrorCode;
+import com.example.board.repository.post.PostRepository;
+import com.example.board.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.NoSuchElementException;
 
 @Service
 @Transactional
@@ -26,7 +26,7 @@ public class PostService {
     private final UserRepository userRepository;
 
     public Long createPost(CreatePostRequest requestDto) {
-        final User user = userRepository.findById(requestDto.authorId()).orElseThrow(() -> new NoSuchElementException("유저가 없음"));
+        final User user = userRepository.findById(requestDto.authorId()).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         final Post post = postRepository.save(PostConverter.toPost(requestDto, user));
         return post.getId();
     }
@@ -37,17 +37,17 @@ public class PostService {
 
     @Transactional(readOnly = true)
     public PostResponse getPost(Long id) {
-        final Post post = postRepository.findById(id).orElseThrow(() -> new NoSuchElementException("게시글이 없음"));
+        final Post post = postRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
         return PostConverter.toPostResponse(post);
     }
 
     public void updatePost(Long id, UpdatePostRequest requestDto) {
-        final Post post = postRepository.findById(id).orElseThrow(() -> new NoSuchElementException("게시글이 없음"));
+        final Post post = postRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
         post.update(requestDto.title(), requestDto.content());
     }
 
     public void deletePost(Long id) {
-        final Post post = postRepository.findById(id).orElseThrow(() -> new NoSuchElementException("게시글이 없음"));
+        final Post post = postRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
         postRepository.delete(post);
     }
 }
