+++ src/main/java/com/example/board/service/UserService.java
@@ -5,12 +5,12 @@ import com.example.board.domain.User;
 import com.example.board.dto.request.CreateUserRequest;
 import com.example.board.dto.request.UpdateUserRequest;
 import com.example.board.dto.response.UserResponse;
-import com.example.board.repository.UserRepository;
+import com.example.board.exception.CustomException;
+import com.example.board.exception.ErrorCode;
+import com.example.board.repository.user.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.NoSuchElementException;
 
 @Service
 @Transactional
@@ -21,24 +21,27 @@ public class UserService {
 
     public Long createUser(CreateUserRequest requestDto) {
         if (userRepository.existsByName(requestDto.name())) {
-            throw new IllegalArgumentException("이미 존재하는 이름입니다.");
+            throw new CustomException(ErrorCode.DUPLICATED_USER_NAME);
         }
         return userRepository.save(UserConverter.toUser(requestDto)).getId();
     }
 
     @Transactional(readOnly = true)
     public UserResponse getUser(Long id) {
-        final User user = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("유저가 존재하지 않습니다."));
+        final User user = userRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         return UserConverter.toUserResponse(user);
     }
 
     public void updateUser(Long id, UpdateUserRequest requestDto) {
-        final User user = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("유저가 존재하지 않습니다."));
+        final User user = userRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         user.update(requestDto.name(), requestDto.age(), requestDto.hobby());
     }
 
     public void deleteUser(Long id) {
-        final User user = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("유저가 존재하지 않습니다."));
-        userRepository.delete(user);
+        final User user = userRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+        if (user.isDeleted()) {
+            throw new CustomException(ErrorCode.ALREADY_DELETED_USER);
+        }
+        user.delete();
     }
 }
